Add JSDocs
