+++ src/core/events/KeyDownEvent.ts
@@ -1,4 +1,7 @@
 import { DomEvent } from './DomEvent';
 
 
+/**
+ * Key down DOM event
+ */
 export type KeyDownEvent = DomEvent<'keyDown', KeyboardEvent>;
