+++ src/core/events/LoadEvent.ts
@@ -1,4 +1,7 @@
 import { DomEvent } from './DomEvent';
 
 
+/**
+ * Window load DOM event
+ */
 export type LoadEvent = DomEvent<'load', Event>;
