+++ src/core/events/ResizeEvent.ts
@@ -1,4 +1,7 @@
 import { DomEvent } from './DomEvent';
 
 
+/**
+ * Window resize DOM event
+ */
 export type ResizeEvent = DomEvent<'resize', UIEvent>;
