+++ src/core/handlers/CanvasClearEventHandler.ts
@@ -2,6 +2,9 @@ import { EventHandler } from '../events/EventBus';
 import { State } from '../../game/State';
 
 
+/**
+ * Event handler that clears canvas every tick
+ */
 export const CanvasClearEventHandler: EventHandler<State> = ({ canvas, context }, { type }) => {
     if (type === 'tick') {
         context.clearRect(0, 0, canvas.width, canvas.height);
