+++ src/core/handlers/InitEventHandler.ts
@@ -2,6 +2,9 @@ import { EventHandler } from '../events/EventBus';
 import { State } from '../../game/State';
 
 
+/**
+ * Event handler that updates canvas size on window load and window resize
+ */
 export const InitEventHandler: EventHandler<State> = ({ canvas }, event) => {
     if (event.type !== 'load' && event.type != 'resize') {
         return;
