+++ src/game/Collider.ts
@@ -1,12 +1,17 @@
-import { Coordinates } from './Coordinates';
+import { Coordinates } from '../utils/Coordinates';
 
 
+/**
+ * Collider object, could check is collides with point in world
+ */
 export interface Collider {
 
     /**
      * Check is object collides with point in world
      *
-     * @param point
+     * @param point point to check
+     *
+     * @return `true` if collides, `false` otherwise
      */
     collides(point: Coordinates): boolean;
 }
