+++ src/game/GameWorld.ts
@@ -1,20 +1,46 @@
-import { Coordinates } from './Coordinates';
-import { Player } from './Player/Player';
-import { Drawable } from './Drawable';
+import { Coordinates } from '../utils/Coordinates';
 import { GameObject } from './GameObject';
+import { Player } from './Player/Player';
 import { isTrigger } from './Trigger';
+import { Drawable } from './Drawable';
 
 
+/**
+ * Game world
+ *
+ * Holds objects that placed in world and player
+ *
+ * Implements base game logic
+ */
 export class GameWorld implements Drawable {
 
+    /**
+     * Array of game objects, placed in world
+     *
+     * By default, only player
+     *
+     * Order of objects controls drawing order (firstly first objects, at end last)
+     *
+     * Must contain player (same as {@link player})
+     */
     objects: GameObject[] = [];
 
+    /**
+     * Player
+     */
     player: Player = new Player();
 
     constructor() {
         this.objects.push(this.player);
     }
 
+    /**
+     * Place new object in world
+     *
+     * Places object strictly on top under player
+     *
+     * @param object object to place
+     */
     placeObject(object: GameObject) {
         // FIXME dirty hack to select correct place for object,
         //       where it will be on top but under player and walls
@@ -29,6 +55,11 @@ export class GameWorld implements Drawable {
         this.objects.splice(playerIdx, 0, object);
     }
 
+    /**
+     * Removes object from world
+     *
+     * @param object object to remove
+     */
     removeObject(object: GameObject): boolean {
         const idx = this.objects.indexOf(object);
 
@@ -40,6 +71,15 @@ export class GameWorld implements Drawable {
         return true;
     }
 
+    /**
+     * Move player by specified vector
+     *
+     * Checks can player step on result cell and runs triggers
+     *
+     * @param direction vector to add to player's coordinates
+     *
+     * @return `true` if player was moved, `false` if there is impassable object on cell
+     */
     movePlayer(direction: Coordinates): boolean {
         const newPosition = new Coordinates(
             this.player.coordinates.x + direction.x,
