+++ src/game/generators/BoxedWorldGenerator.ts
@@ -1,13 +1,13 @@
 import { Random } from 'random';
 
 import { ItemGenerator, WorldGenerator } from './Generator';
-import { Coordinates } from '../Coordinates';
-import { GameWorld } from '../GameWorld';
-import { sample } from '../../utils/shuffle';
 import { delayPromise } from '../../utils/delayPromise';
+import { Coordinates } from '../../utils/Coordinates';
+import { DroppedItem } from '../DroppedItem';
+import { sample } from '../../utils/shuffle';
 import { GameObject } from '../GameObject';
 import { range } from '../../utils/range';
-import { DroppedItem } from '../DroppedItem';
+import { GameWorld } from '../GameWorld';
 import { Wall } from '../Wall/Wall';
 
 
@@ -22,14 +22,27 @@ class Box {
     }
 }
 
+/**
+ * World generator that generates world as closed area,
+ * consisting of boxes with same size (like in The binding of Isaac)
+ *
+ * Each box has random number of doors to other boxes and random number of random items,
+ * generated by specified item generator
+ */
 export class BoxedWorldGenerator implements WorldGenerator {
 
-    readonly maxDepth: number;
-    readonly maxItemsInRoom: number;
-    readonly boxSize: Coordinates;
+    private readonly maxDepth: number;
+    private readonly maxItemsInRoom: number;
+    private readonly boxSize: Coordinates;
 
-    readonly itemGenerator: ItemGenerator;
+    private readonly itemGenerator: ItemGenerator;
 
+    /**
+     * @param maxDepth max depth of generating tree
+     * @param maxItemsInRoom max amount of items in each room
+     * @param boxSize box size (width and height)
+     * @param itemGenerator item generator
+     */
     constructor(maxDepth: number, maxItemsInRoom: number, boxSize: Coordinates, itemGenerator: ItemGenerator) {
         this.maxDepth = maxDepth;
         this.maxItemsInRoom = maxItemsInRoom;
