+++ src/game/generators/UniformItemGenerator.ts
@@ -4,13 +4,25 @@ import { ItemGenerator } from './Generator';
 import { Item } from '../items/Item';
 
 
+/**
+ * Pair of item generator and (relative) probability
+ */
 export type ItemProbabilityPair = [ItemGenerator, number];
 
+/**
+ * Item generator that generates items with specified relative probability using
+ * specified item generators
+ *
+ * Real probability of each item is its relative probability divided by sum of all relative probabilities
+ */
 export class UniformItemGenerator implements ItemGenerator {
 
-    readonly items: readonly ItemProbabilityPair[];
+    private readonly items: readonly ItemProbabilityPair[];
     private readonly sumProbability: number;
 
+    /**
+     * @param items array of item generators with relative probabilities
+     */
     constructor(items: ItemProbabilityPair[]) {
         this.items = [...items];
