+++ src/game/items/Item.ts
@@ -2,6 +2,9 @@ import { GameWorld } from '../GameWorld';
 import { Drawable } from '../Drawable';
 
 
+/**
+ * Item, could be used by player
+ */
 export interface Item extends Drawable {
 
     /**
@@ -9,7 +12,7 @@ export interface Item extends Drawable {
      *
      * @param world game world
      *
-     * @return what item must be placed instead of this
+     * @return what item must be placed to inventory instead of this
      */
     useItem(world: GameWorld): Item | undefined;
 }
