+++ src/game/items/TexturedItem.ts
@@ -1,13 +1,26 @@
 import { drawTexture } from '../../utils/drawTexture';
-import { Coordinates } from '../Coordinates';
+import { Coordinates } from '../../utils/Coordinates';
 import { GameWorld } from '../GameWorld';
 import { Item } from './Item';
 
 
+/**
+ * Abstract class that provides convenient way to texture items
+ */
 export abstract class TexturedItem implements Item {
 
+    /**
+     * Item texture
+     *
+     * @protected
+     */
     protected readonly texture: string;
 
+    /**
+     * @param texture item texture
+     *
+     * @protected
+     */
     protected constructor(texture: string) {
         this.texture = texture;
     }
