+++ src/index.ts
@@ -11,8 +11,8 @@ import { FileWorldGenerator } from './game/generators/FileWorldGenerator';
 import { GoldenApple } from './game/items/GoldenApple/GoldenApple';
 import { constGenerator } from './game/generators/Generator';
 import { seededRandom } from './utils/seededRandom';
+import { Coordinates } from './utils/Coordinates';
 import { DroppedItem } from './game/DroppedItem';
-import { Coordinates } from './game/Coordinates';
 import { Sword } from './game/items/Sword/Sword';
 import { GameWorld } from './game/GameWorld';
 import { Wall } from './game/Wall/Wall';
