+++ src/utils/drawTexture.ts
@@ -1,8 +1,13 @@
-import { Coordinates } from '../game/Coordinates';
+import { Coordinates } from './Coordinates';
 
 
 const textures = new Map();
 
+/**
+ * Load texture to use in {@link drawTexture}
+ *
+ * @param texture imported texture
+ */
 export const loadTexture = (texture: string) => {
     const img = new Image();
     img.src = texture;
@@ -10,6 +15,18 @@ export const loadTexture = (texture: string) => {
     textures.set(texture, img);
 }
 
+/**
+ * Draw loaded texture on canvas
+ *
+ * This is an error if texture wasn't loaded before call
+ *
+ * @param texture texture to draw
+ * @param context context to draw on
+ * @param place center of texture in pixels
+ * @param scale size of texture in pixels (resolution of texture isn't preserved)
+ *
+ * @see loadTexture
+ */
 export const drawTexture = (
     texture: string,
     context: CanvasRenderingContext2D,
